+++ VERSION
@@ -1,1 +1,1 @@
-1.1.1
+1.2.0
