--- tracker-card-editor.js
@@ -1,98 +1,0 @@
-if (!customElements.get("paper-input")) {
-  console.log("imported", "paper-input");
-  import("https://unpkg.com/@polymer/paper-input/paper-input.js?module");
-}
-
-const fireEvent = (node, type, detail, options) => {
-  options = options || {};
-  detail = detail === null || detail === undefined ? {} : detail;
-  const event = new Event(type, {
-    bubbles: options.bubbles === undefined ? true : options.bubbles,
-    cancelable: Boolean(options.cancelable),
-    composed: options.composed === undefined ? true : options.composed
-  });
-  event.detail = detail;
-  node.dispatchEvent(event);
-  return event;
-};
-
-const LitElement = Object.getPrototypeOf(
-  customElements.get("ha-panel-lovelace")
-);
-const html = LitElement.prototype.html;
-
-export class TrackerCardEditor extends LitElement {
-  setConfig(config) {
-    this._config = config;
-  }
-
-  static get properties() {
-    return { hass: {}, _config: {} };
-  }
-
-  get _title() {
-    return this._config.title || "";
-  }
-
-  render() {
-    if (!this.hass) {
-      return html``;
-    }
-
-    return html`
-      ${this.renderStyle()}
-      <div class="card-config">
-        <div class="side-by-side">
-          <paper-input
-            label="Title (Optional)"
-            .value="${this._title}"
-            .configValue="${"title"}"
-            @value-changed="${this._valueChanged}"
-          ></paper-input>
-        </div>
-        <div>For more advanced configuration use the yaml editor</div>
-      </div>
-    `;
-  }
-
-  renderStyle() {
-    return html`
-      <style>
-        paper-toggle-button {
-          padding-top: 16px;
-        }
-        .side-by-side {
-          display: flex;
-        }
-        .side-by-side > * {
-          flex: 1;
-          padding-right: 4px;
-        }
-      </style>
-    `;
-  }
-
-  _valueChanged(ev) {
-    if (!this._config || !this.hass) {
-      return;
-    }
-    const target = ev.target;
-    if (this[`_${target.configValue}`] === target.value) {
-      return;
-    }
-    if (target.configValue) {
-      if (target.value === "") {
-        delete this._config[target.configValue];
-      } else {
-        this._config = {
-          ...this._config,
-          [target.configValue]:
-            target.checked !== undefined ? target.checked : target.value
-        };
-      }
-    }
-    fireEvent(this, "config-changed", { config: this._config });
-  }
-}
-
-customElements.define("tracker-card-editor", TrackerCardEditor);
